Payment Service added

+++ end-to-end-tests/src/test/java/behaviourtests/AccountRegistrationService.java
@@ -7,16 +7,39 @@ import javax.ws.rs.client.WebTarget;
 
 public class AccountRegistrationService {
 
-	public Account register(Account c) {
+	public Account register(Account c) throws AccountAlreadyExists {
 		Client client = ClientBuilder.newClient();
 		WebTarget r = client.target("http://localhost:8080/");
-		var response = r.path("accounts").request().post(Entity.json(c), Account.class);
-		return response;
+
+		var response = r.path("accounts").request().post(Entity.entity(c, "application/json"));
+
+		switch(response.getStatus()) {
+			case 200:
+				return response.readEntity(Account.class);
+			case 409:
+				throw new AccountAlreadyExists("that account already exists");
+			default:
+				return null;
+		}
 	}
 
 	public void deleteAccount(String accountId) {
+		if(accountId==null){
+			return;
+		}
 		Client client = ClientBuilder.newClient();
 		WebTarget r = client.target("http://localhost:8080/");
 		r.path("accounts").path(accountId).request().delete();
+	}
+
+	public Account getAccount(String accountId) throws NoSuchAccountException {
+		Client client = ClientBuilder.newClient();
+		WebTarget r = client.target("http://localhost:8080/");
+		var response = r.path("accounts").path(accountId).request().get();
+		if(response.getStatus() == 200) {
+			return response.readEntity(Account.class);
+		} else {
+			throw new NoSuchAccountException("Account doesn't exist");
+		}
 	}
 }
